+++ src/models/Notification.cs
@@ -1,13 +1,14 @@
 // Copyright (C) 2023 Karl Pickett / Vilark Project
 namespace vilark;
 
+record struct LoadingProgress(int Processed, int Ignored);
+
 // A grab-bag of message types that we can send to the main thread
 record struct Notification(
-        int? Processed = null,
-        int? Ignored = null,
+        LoadingProgress? LoadingProgress = null,
         IEnumerable<ISelectableItem>? CompletedData = null,
-        string? ErrorMessage = null,
+        string? FatalErrorMessage = null,
         string? WebRequest = null,
         bool ForceRedraw = false,
         bool ChildExited = false
-        );
+    );