+++ src/util/UnixProcess.cs
@@ -6,29 +6,32 @@ using System.Runtime.InteropServices;
 namespace vilark;
 
 /*
- * When exec'ing a child (replacement) process, these things need to be considered:
+ * When starting a replacement (execve) or child (Process.Start) process, these
+ * things need to be considered:
  *
- * 1. Putting the terminal back to the 'main' screen.  Even if this is only for a moment,
- *    the convention is that each process leaves the terminal as it was when they started.
- *    Whatever we leave it at, is what the next process will leave it at, by default.
+ * 1. How the terminal will be set to the alternate screen, and more importantly, if it
+ *    will be returned to the main screen when the process exits.
  *
- *    Note the vim docs:  "When Vim exits, the terminal will be put back into
- *    the mode it was before Vim started."
+ *    Vim can stop the usual "put screen back to main on exit" convention if
+ *    we set t_ti and t_te to empty at startup, see ":help restorescreen"
  *
  * 2. Resetting all terminal colors/underlining back to the default, and enabling the cursor.
+ *    This can be done easily just by write()ing escape codes to /dev/tty.
  *
  * 3. Setting the mode of /dev/tty back to cooked/echo, or at least how it was when you started.
  *    This needs an ioctl(), sadly.  Or, you hack it and run 'stty sane' in a subprocess.
  *
+ *    From the vim docs:  "When Vim exits, the terminal will be put back into
+ *    the mode it was before Vim started."
+ *
  * 4. Restoring signal masks to the default.  execve() does this for you, otherwise
- *    if you are using .NET Process.Start(), you should unregister your signal handlers first.
+ *    if you are using .NET Process.Start(), perhaps unregister your signal handlers first?
  *
- * 5. Making sure nothing else is still reading from /dev/tty.  Again, execve() does this for you,
- *    because all threads are terminated.  The .NET ReadKey() API is awkward to cancel otherwise.
+ * 5. Making sure nothing else is still reading from /dev/tty.  execve() does this for you,
+ *    because all threads are terminated.  If starting a child with
+ *    Process.Start(), ensure no other thread is calling ReadKey().
  *
- * Note that 1) and 2) can be done easily just by write()ing escape codes to /dev/tty.
- *
- * After trying a lot of failed experiments, the most robust solution for 3-5 is to
+ * After a lot of failed experiments, the most robust solution for 3-5 is to
  * execve() a wrapper shell command that calls "stty sane" and "exec $target_command"
  * in one shot.  Note that calling them separately was not always working, perhaps because
  * some .NET Console code was being triggered by child processes starting/exiting.
@@ -83,9 +86,9 @@ class UnixProcess
     }
 
     // Use the standard .Net runtime to launch a child process.
-    // When it finishes, send an event to the loadingEvent queue.
+    // When it finishes, send an event to the notifications queue.
     static public void StartChild(string program, string[] args,
-                    EventQueue<Notification> loadingEvent) {
+                    EventQueue<Notification> notifications) {
         //UnregisterSignals();
         //RunTerminalReset();
         var pi = new ProcessStartInfo(program);
@@ -95,7 +98,7 @@ class UnixProcess
             Task.Run(() => {
                     p.WaitForExit();
                     Log.Info($"exec process finished: {p.ExitCode}");
-                    loadingEvent.AddEvent(new Notification(ChildExited: true));
+                    notifications.AddEvent(new Notification(ChildExited: true));
                 });
         } else {
             Log.Info("exec process didn't start");