+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Vilark Changelog
 
+## 2023.9.13 (Version 1.6.2)
+* Small refactor of startup code
+* Forward exceptions to main thread
+* Don't allow duplicate SIGCONT/SIGTSTP signals
+
 ## 2023.9.12 (Version 1.6.1)
 * Fix flickering on some terminals
 * Make child vim and vilark proceses not mess with the terminal,